+++ modules/miscellaneous.js
@@ -109,10 +109,9 @@ async function fetchAllChannelMessages(channel, limit) {
 
 exports.remind = async (args, user) => {
     if (args.length < 3) return `Command usage: ${config.prefix}remind <time> <message>`
-    const ms = require('ms');
-    const timeToMs = require('./timetoms');
-    time = await timeToMs(args[1])
+    const timeConvert = require('./timeconvert');
+    time = await timeConvert(args[1], { convertTo: 'ms' })
     if (!parseInt(time)) return `That doesn't look like a valid time`;
     await dataManager.writeData('reminders', { user: user.id, reminder: args.slice(2).join(' '), time: Date.now() + time }, { user: user.id, reminder: args.slice(2).join(' '), time: Date.now() + time })
-    return `I'll remind you in ${args[1]}:\n\`${args.slice(2).join(' ')}\``;
+    return `I'll remind you in ${await timeConvert(args[1], { convertTo: 'long' })}:\n\`${args.slice(2).join(' ')}\``;
 }