+++ modules/schemas/mutes-schema.js
@@ -1,0 +1,10 @@
+const mongoose = require('mongoose');
+
+const muteSchema = mongoose.Schema({
+    guild: String,
+    user: String,
+    expires: Number,
+    unmuteAttempts: Number
+})
+
+module.exports = mongoose.model('mutes', muteSchema)